Lots of work done
Worked on filling in some of the functions in the HostConnection class,
ClientConnection, as well as added some new fields into the
ServerStateSaver ClinetStateSaver objects. Still have to finish all
connection algorithms then build the connection handling for client and
server. Then build a logging class to handle logging while sending and
receiving.