Built ResponseBuilder....
Added to ProcessMessage so that it calls ResponseBuilder and processes
message to prepare for sending. Changed ClientConnection to use the
ServerStateSaver object instead of just ClientStateSaver. This enabled
the ResponseBuilder to update send message and allows the
ClientConnection to send formatted message to the endpoint server. Still
have to step through algorithm to be sure that it will work properly
when it does run.