Created new project under same repository for middleware
Since this is for the same purpose, just new project, I just created the
project in the same repository folder. Have a server running to
manipulate messages and parrot them back to client. Have to build
"client" side to handle endpoint communication. From working server to
editing the "send function" in this to send to endpoint, receive and
send back to client to finish.